Improving coverage and tests

+++ mqtt-km/src/commonMain/kotlin/io/github/petretiandrea/mqtt/core/model/packets/PubAck.kt
@@ -25,7 +25,7 @@ data class PubAck(
 
     companion object : MqttDeserializer {
         override fun fromByteArray(data: UByteArray): Result<PubAck> {
-            val messageId = Util.getIntFromMSBLSB(data[0].toByte(), data[1].toByte())
+            val messageId = Util.getIntFromMSBLSB(data[0], data[1])
             return Result.success(PubAck(messageId))
         }
     }
