+++ mqtt-km/src/commonMain/kotlin/io/github/petretiandrea/mqtt/core/model/packets/PubComp.kt
@@ -25,7 +25,7 @@ data class PubComp(
 
     companion object : MqttDeserializer {
         override fun fromByteArray(data: UByteArray): Result<PubComp> {
-            val messageId = Util.getIntFromMSBLSB(data[0].toByte(), data[1].toByte())
+            val messageId = Util.getIntFromMSBLSB(data[0], data[1])
             return Result.success(PubComp(messageId))
         }
     }
