+++ mqtt-km/src/commonMain/kotlin/io/github/petretiandrea/mqtt/core/model/packets/PubRec.kt
@@ -25,7 +25,7 @@ data class PubRec(
 
     companion object : MqttDeserializer {
         override fun fromByteArray(data: UByteArray): Result<PubRec> {
-            val messageId = Util.getIntFromMSBLSB(data[0].toByte(), data[1].toByte())
+            val messageId = Util.getIntFromMSBLSB(data[0], data[1])
             return Result.success(PubRec(messageId))
         }
     }
